Convert BitBoardScorer to use int instead of double.

+++ src/main/java/ut/veikotiit/checkers/scorer/BitBoardScorer.java
@@ -4,5 +4,5 @@ import ut.veikotiit.checkers.Color;
 import ut.veikotiit.checkers.bitboard.BitBoard;
 
 public interface BitBoardScorer {
-    double getScore(BitBoard bitBoard, Color color);
+    int getScore(BitBoard bitBoard, Color color);
 }
